fix: rename decideversionstatus to getapplicationupdatestatus
- rename helper function and test file
- fix jsdoc
